agregando funcionalidad a la persistencia hay un error al guardar la molotov

+++ trunk/BombitaTp2/src/modelo/armamento/Armamentable.java
@@ -4,5 +4,5 @@ public interface Armamentable {
 
 	public float obtenerTime();
 	
-	public void cambiarTime(float Tiempo);
+	public void cambiarTime(int Tiempo);
 }
